+++ crates/api/src/lib.rs
@@ -1,12 +1,14 @@
 mod library;
 #[cfg(feature = "serde")]
 mod manifest;
+mod project;
 mod records;
 mod registry;
 
 pub use library::*;
 #[cfg(feature = "serde")]
 pub use manifest::*;
+pub use project::*;
 pub use records::*;
 pub use registry::*;