+++ crates/api/src/library.rs
@@ -4,6 +4,7 @@ use shared::IdVec;
 
 use crate::{ApiAdt, ApiConstant, ApiEntry, ApiFunction, ApiMethod, Intrinsic, NativeId, Registry};
 
+#[derive(Clone)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[cfg_attr(feature = "serde", serde(bound(deserialize = "C: Default")))]
 pub struct Library<C> {